feat: linked playback behaviors with command and controller
