+++ src/messages.json
@@ -8,8 +8,9 @@
 	"Command.Play.NotAllowed": "You're not allowed to use the bot. It may be because you're not in a voice channel or not in the same channel as the bot.",
 	"Command.Play.InvalidLink": "This link is invalid. Try another one.",
 	"Command.Play.NotFound": "Video doesn't exist",
-	"Command.Play.GenericError": "Could't find a video matching this query. It might be restricted, or the search quota might have been depleted. If you're sure this video exists try using a link instead.",
-	"Command.Play.Queued": "Queued [title](url) [member]",
+	"Command.Play.GenericError": "Could't find a video matching this query. It might be restricted or private. If you're sure this video exists try using a different link.",
+	"Command.Play.Playlist": "Queued number tracks. Use `queue` command to view the full list",
+	"Command.Play.Single": "Queued [title](url) [member]",
 	"Command.Play.NotPaused": "The bot is not paused.",
 	"Command.Play.AlreadyPlaying": "The bot is not paused",
 	"Command.Play.Resumed": "Resumed ️▶️",
