+++ src/utils/general.ts
@@ -206,3 +206,15 @@ export const formatDuration = (duration: number) =>
 		colonNotation: true,
 		secondsDecimalDigits: 0,
 	});
+
+/**
+ * Parses a time string HH:MM:SS into milliseconds
+ */
+export const stringToTimestamp = (time: string) => {
+	if (!time) return undefined;
+
+	return time
+		.split(":")
+		.map(period => Number(period))
+		.reduce((accumulator, period) => 60 * accumulator + period, 0);
+};
